requirements.txt
# requirements.txt
openai
psycopg2-binary
sqlalchemy
chromadb
sentence-transformers
python-dotenv
langchain
python-frontmatter
celery
redis